+++ web/apps/photos/src/pages/shared-albums.tsx
@@ -9,6 +9,7 @@ import {
 } from "@/new/photos/components/Gallery/ListHeader";
 import { SpaceBetweenFlex } from "@/new/photos/components/mui";
 import downloadManager from "@/new/photos/services/download";
+import { sortFiles } from "@/new/photos/services/files";
 import { EnteFile } from "@/new/photos/types/file";
 import { mergeMetadata } from "@/new/photos/utils/file";
 import {
@@ -76,7 +77,7 @@ import {
     downloadCollectionFiles,
     isHiddenCollection,
 } from "utils/collection";
-import { downloadSelectedFiles, getSelectedFiles, sortFiles } from "utils/file";
+import { downloadSelectedFiles, getSelectedFiles } from "utils/file";
 import { formatNumber } from "utils/number/format";
 import { PublicCollectionGalleryContext } from "utils/publicCollectionGallery";
