+++ web/packages/new/photos/services/files.ts
@@ -38,6 +38,22 @@ export const setLocalFiles = async (
     await localForage.setItem(tableName, files);
 };
 
+export function sortFiles(files: EnteFile[], sortAsc = false) {
+    // sort based on the time of creation time of the file,
+    // for files with same creation time, sort based on the time of last modification
+    const factor = sortAsc ? -1 : 1;
+    return files.sort((a, b) => {
+        if (a.metadata.creationTime === b.metadata.creationTime) {
+            return (
+                factor *
+                (b.metadata.modificationTime - a.metadata.modificationTime)
+            );
+        }
+        return factor * (b.metadata.creationTime - a.metadata.creationTime);
+    });
+}
+
+
 export const TRASH = "file-trash";
 
 export async function getLocalTrash() {
